commit vymaz DAO

+++ Ekniznica/src/main/java/sk/paz1c/Elibrary/interfaces/RentedBookDao.java
@@ -11,6 +11,6 @@ public interface RentedBookDao {
 	public List<RentedBook> getAllRenteBooksForAdmin();
 	public RentedBook returnBook(RentedBook rentedBook);
 	public List<RentedBook> getNonReturnedRentedBookById(long id);
-	public void addRentedBook(int userId, int bookId);
+	public void addRentedBook(Long userId, Long bookId);
 	
 }
